Move Endpoint, Ip and Port to separate modules
